Troubleshooted stack view
The top picure was not always the largest picture on the screen. It took a very long time to sort this issue out. I fixed this by making the big pic >= 70, smaller pics <= 70, and set the text constraints to = 20. Whenever the app is in portrait mode, the buttons are aligned vertically, but if landscape then it's horizontal.